Fixes to fit posting views to data structure
Data model continuity fixes
